--- QuantaTextToSql/__init__.py
@@ -1,7 +1,0 @@
-# pyright: reportUnusedImport=false
-
-from QuantaTextToSql.ablate import *
-from QuantaTextToSql.corrupt_data import *
-from QuantaTextToSql.finetune import *
-from QuantaTextToSql.load_data import *
-from QuantaTextToSql.training_data import *
