+++ TinySQL/finetune/delete_from_hf.py
