+++ TinySQL/finetune/run_experiments.sh
