+++ TinySQL/training_data/fragments/english_aggregates.py
