+++ TinySQL/training_data/fragments/english_select_from.py
