+++ TinySQL/training_data/fragments/table_names.py
