+++ TinySQL/training_data/generate_cs1.py
