+++ TinySQL/training_data/generate_cs2.py
