+++ TinySQL/training_data/generate_cs3.py
