+++ TinySQL/training_data/generate_csn.py
