+++ TinySQL/training_data/sql_create_table.py
