+++ TinySQL/training_data/sql_select_from.py
